Move relations to app/
This is what we expect to do for the eventual integration of rom-rb into Hanami.
