+++ config/providers/persistence.rb
@@ -24,8 +24,8 @@ Hanami.app.register_provider :persistence, namespace: true do
   start do
     rom_config = target["persistence.config"]
     rom_config.auto_registration(
-      target.root.join("lib/decafsucks/persistence"),
-      namespace: "Decafsucks::Persistence"
+      target.root.join("app"),
+      namespace: Hanami.app.namespace.to_s
     )
 
     register "rom", ROM.container(rom_config)
