--- lib/decafsucks/persistence/relations/users.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-module Decafsucks
-  module Persistence
-    module Relations
-      class Users < ROM::Relation[:sql]
-        schema :users, infer: true
-      end
-    end
-  end
-end
